pyproject.toml
[tool.poetry]
name = "mev-boost-relay-allowed-list"
version = "0.1.0"
description = "Contract storing a list of MEV boost relays approved by Lido DAO"
authors = ["Lido dev team"]
license = "MIT"
readme = "README.md"
homepage = "https://mainnet.lido.fi"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
eth-ape = "~0.6.22"

[tool.poetry.dev-dependencies]
black = "~22.6.0"
pre-commit = "~2.19.0"
vyper = "0.3.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
